requirements.txt
openai
psycopg2-binary
fastapi
pydantic
